[Chore] #139 리뷰 조회 관련 Entity, DTO model 파일 분리
